translate files moved to lib and converted to ts
