+++ packages/modal-ui-js/src/lib/components/WalletConnectionFailed.ts
@@ -1,7 +1,7 @@
 import { ModuleState, Wallet } from "@near-wallet-selector/core";
 import { connectToWallet } from "../render-modal";
 // @ts-ignore
-import { translate } from "../../../translate/translate";
+import { translate } from "../translate/translate";
 
 export async function renderWalletConnectionFailed(
   module: ModuleState<Wallet>,
