+++ packages/modal-ui-js/src/lib/components/WalletNotInstalled.ts
@@ -2,7 +2,7 @@ import { ModuleState, Wallet } from "@near-wallet-selector/core";
 import { modalState } from "../modal";
 
 // @ts-ignore
-import { translate } from "../../../translate/translate";
+import { translate } from "../translate/translate";
 
 export function renderWalletNotInstalled(
   module: ModuleState<Wallet> & {
