+++ packages/modal-ui-js/src/lib/locale/de.json
