+++ packages/modal-ui-js/src/lib/locale/es.json
