+++ packages/modal-ui-js/src/lib/locale/fr.json
